Request all open tickets from a Trac install at once.
This prevents repeated HTTP requests for individual tickets and
allows us to keep knownWPBug (or @ticket - see #102) in place for
tickets already closed, without any real performance penalty.
This is good for tracking and also in case a bug is reopened.

This shaved 30 seconds off the test suite run time for me.



git-svn-id: https://unit-test.svn.wordpress.org/trunk@891 4a5a2adb-35eb-4268-9a1b-fe04a4134342